+++ weldx/asdf/extension.py
@@ -10,6 +10,12 @@ from .types import _weldx_asdf_types, _weldx_types
 __all__ = ["WeldxExtension", "WeldxAsdfExtension"]
 
 
+class WxSyntaxError(Exception):
+    """Exception raising on custom weldx ASDF syntax errors."""
+
+    pass
+
+
 class WeldxExtension(AsdfExtension):
     """Extension class registering types with both tags and schemas defined by weldx."""