+++ weldx/asdf/validators.py
@@ -6,6 +6,7 @@ import pandas as pd
 from asdf import ValidationError
 from asdf.schema import _type_to_tag
 
+from weldx.asdf.extension import WxSyntaxError
 from weldx.constants import WELDX_QUANTITY as Q_
 from weldx.constants import WELDX_UNIT_REGISTRY as UREG
 
@@ -145,18 +146,26 @@ def _compare(_int, exp_string):
         True or False
     """
     if _int < 0:
-        raise ValueError("Negative dimension found")
+        raise WxSyntaxError("Negative dimension found")
 
     if ":" in exp_string:
         ranges = exp_string.split(":")
 
         if ranges[0] == "":
             ranges[0] = 0
+        elif ranges[0].isnumeric():
+            ranges[0] = int(ranges[0])
+        else:
+            raise WxSyntaxError(f"Non numeric character in range {exp_string}")
         if ranges[1] == "":
             ranges[1] = _int
+        elif ranges[1].isnumeric():
+            ranges[1] = int(ranges[1])
+        else:
+            raise WxSyntaxError(f"Non numeric character in range {exp_string}")
 
-        if int(ranges[0]) > int(ranges[1]):
-            raise ValueError(f"The range should not be descending in {exp_string}")
+        if ranges[0] > ranges[1]:
+            raise WxSyntaxError(f"The range should not be descending in {exp_string}")
         return int(ranges[0]) <= _int <= int(ranges[1])
 
     else:
@@ -250,15 +259,15 @@ def _validate_expected_list(list_expected):
     validator = 0
     for exp in list_expected:
         if validator == 1 and not ("(" in str(exp) or "..." in str(exp)):
-            raise ValueError(
+            raise WxSyntaxError(
                 "Optional dimensions in the expected "
                 "shape should only stand at the end/beginning."
             )
         if validator == 2:
-            raise ValueError('After "..." should not be another dimension.')
+            raise WxSyntaxError('After "..." should not be another dimension.')
         if "..." in str(exp):
             if "..." != exp:
-                raise ValueError(
+                raise WxSyntaxError(
                     f'"..." should not have additional properties:' f" {exp} was found."
                 )
             validator = 2
@@ -269,14 +278,14 @@ def _validate_expected_list(list_expected):
                 or len(val.group(1)) + 2 != len(exp)
                 or not _is_range_format_valid(val.group(1))
             ):
-                raise ValueError(
+                raise WxSyntaxError(
                     f'Invalid optional dimension format. Correct format is "(_)", but '
                     f" {exp} was found."
                 )
 
             validator = 1
         elif not _is_range_format_valid(str(exp)):
-            raise ValueError(
+            raise WxSyntaxError(
                 f"{exp} is an invalid range format."
                 f"Consult the documentation for a list of all valid options"
             )
@@ -449,7 +458,7 @@ def _custom_shape_validator(dict_test, dict_expected):
                 elif dict_values[key] != _dict_values[key]:
                     return False
     else:
-        raise ValueError(
+        raise WxSyntaxError(
             f"Found an incorrect object: {type(dict_expected)}. "
             "Should be a dict or list."
         )
@@ -563,7 +572,7 @@ def _compare_tag_version(instance_tag: str, tagname: str):
             [v[0] == v[1] for v in zip(vnum, instance_tag_version)]
         )
     else:
-        raise ValueError(f"Unknown wx_tag syntax {tagname}")
+        raise WxSyntaxError(f"Unknown wx_tag syntax {tagname}")
 
     if (not instance_tag.startswith(tag_uri)) or (not version_compatible):
         return False